Extra documentation on locking of undead/skeletal horses.
Cancelling a right click interaction event (as occurs when running a command and
clicking on them) does not prevent the player from mounting these horses. So
there is no point in waiting for the player to mount them before setting the
owner; the horse will always be mounted if right-clicked on.