Put all trainable horse-like animals before untrainable ones in the /hlist numbering, so that they have the lowest numeric IDs for use in /hgps.
Trainable horses are more valuable and useful to players, so players are likely
to want to look for them with /hgps more often, and having them on the first
page of /hlist output is a convenience. In addition, untrainable horses are
improved by conventional Minecraft breeding, so they come and go more frequently
and their position within /hlist fluctuates considerably, essentially mandating
that players consult /hlist to find the current index of an untrainable horse
nearly every time they search with /hgps (if they breed llamas).